+++ src/core/device/device.cc
@@ -67,7 +67,7 @@ void Device::CopyDataToFrom(Block* dst, Block* src, size_t nBytes,
 void Device::RepeatDataToFrom(Block* dst, Block* src, size_t nBytes,
                               CopyDirection direct, bool broadcast_flag, 
                               int axis_shape, int shape_outer, int chunk, 
-                              vector<int> repeats, int dst_offset, int src_offset) {
+                              vector<size_t> repeats, int dst_offset, int src_offset) {
   const char *src_data = reinterpret_cast<const char*>(src->data()) + dst_offset;
   char *dst_data = reinterpret_cast<char*>(dst->mutable_data()) + src_offset;
